Removed datasource layer and merged those changes into repositories directly. Moved this into package
